requirements.txt
jsonrpclib-pelix
pika
pymongo
requests
redis
newspaper3k
python-dateutil
tensorflow
pandas